requirements.txt

python-dotenv>=1.0.1
streamlit>=1.38.0
plotly>=5.22.0
kaleido>=0.2.1
polars>=1.6.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
scipy>=1.13.1
statsmodels>=0.14.2
pydantic>=2.8.2
pyyaml>=6.0.2
joblib>=1.4.2
mlflow>=2.16.0
optuna>=3.6.1
openai>=1.51.0
tenacity>=8.5.0
reportlab>=4.2.2
python-pptx>=0.6.23
